add min stack data structure implmentation
in this solution, we use two stacks, one is for the main stack and other
one is for keep track of the minimum number at run time, so if there is
a value smaller than the current top element of the min stack,we push it
to the min stack    so when we keep push to the main stack, if the
element is greater than    the current min, we essentially do not need
to keep track in the min    stack, cause if we pop that element out of
the main stack, it does not    effect to the current min in the min
stack.    if we pop an element from the main stack and it happens to be
the    current min stack, we pop that element in both of the stacks, so
that    the second smallest element in the min stack on the top of the
min    stack.