add evaluate reverse polish notation
in this question, we are using stack data structure to store the most
the order of the numbers. if we see an operator, we pop the first two
element and do the evaluation and then append the new result to the
stack. Keep doing that until we loop the entire list