add daily temperature problem
We use a stack to keep track days that we have not found any future day
that has higher temperature, the elements in the stack should be in
decreasing order. once we found a day that has higher temperature than
the current top of the stack, we pop it out and calculate the difference
in days. The current day could has higher temperature for multiple days
in the stack, so we keep pop out the elements in the stack until we see
the element in stack has higher temperature than the current day and we
stop. Once we done with the stack checking, we push the current day temp
to the stack and keep doing that to the end of the temperature list