add largest rectangel in histogram
We keep a stack in increasing order,

Keep looping and append to the stack if the next time is higher than the
top item of the stack.
once we see item that is lower than the top item of the the stack
we start poping the top item until the top item is lower than the
current item.
During poping out the top item, we take this opportunity to calculate
the histogram of those area
Once we done with the looping, we got a stack in increasing order, and
start calculate the rest of the historram area
 NOTE: we use the trick is put the first item in the stack with height =
-1 and position = -1